tx2/models/multeye/requirements.txt
absl-py==0.12.0
astor==0.8.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
construct==2.8.8
cycler==0.10.0
dm-tree==0.1.5
dronekit==2.9.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hexdump==3.3
idna==2.10
importlib-metadata==3.7.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
logutils==0.3.5
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.3.4
mock==4.0.3
monotonic==1.6
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pbr==5.5.1
Pillow==7.1.2
pkg-resources==0.0.0
portpicker==1.3.1
protobuf==3.15.6
psutil==5.8.0
ptpy @ file:///home/jetsontx2/Documents/models/multeye/multeyevenv/lib/python3.6/sequoia-ptpy-master
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymavlink==2.4.15
pyparsing==2.4.7
python-dateutil==2.8.1
pyusb==1.1.1
PyYAML==5.4.1
rainbow-logging-handler==2.2.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
#tensorflow==2.4.0+nv21.2
tensorflow-estimator==2.4.0
termcolor==1.1.0
testresources==2.0.1
#tf-model-optimization-nightly==0.5.0.dev20210316
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
